requirements.txt
Pillow==9.2.0
PyYAML==6.0